requirements.txt
configparser
queues
requests
urllib3
requests
ascii_colors
fastapi 
sqlalchemy 
uvicorn 
passlib[bcrypt] 
python-jose[cryptography]
